Fix linter issues
Implement get params

Fix linter issues

+++ philo/src/time.c
@@ -6,12 +6,12 @@
 /*   By: aroque <aroque@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/24 22:11:55 by aroque            #+#    #+#             */
-/*   Updated: 2021/07/13 23:47:58 by aroque           ###   ########.fr       */
+/*   Updated: 2021/07/23 23:55:24 by aroque           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <sys/time.h>
 #include <stdlib.h>
+#include <sys/time.h>
 #include <sys/types.h>
 
 long	timestamp(void)
@@ -19,5 +19,5 @@ long	timestamp(void)
 	struct timeval	now;
 
 	gettimeofday(&now, NULL);
-	return (now.tv_sec * (u_int64_t)1000  + now.tv_usec / 1000);
+	return (now.tv_sec * 1000 + now.tv_usec / 1000);
 }
